requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.3.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
crispy-tailwind==0.5.0
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
kombu==5.2.4
mypy-extensions==0.4.3
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.1
redis==4.2.1
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.3
sentry-sdk==1.5.8
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.8.0
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.0
wsproto==1.1.0
